feat: module8-task1

+++ src/modules/offer/offer.controller.ts
@@ -1,4 +1,5 @@
 import {instanceToPlain} from 'class-transformer';
+import {Types} from 'mongoose';
 import {type Request, type Response} from 'express';
 import {inject, injectable} from 'inversify';
 import {StatusCodes} from 'http-status-codes';
@@ -8,15 +9,15 @@ import {CreateOfferDto} from './dto/create-offer.dto.js';
 import {UpdateOfferDto} from './dto/update-offer.dto.js';
 import {OfferDetailRdo} from './rdo/offer-detail.rdo.js';
 import {OfferSummaryRdo} from './rdo/offer-summary.rdo.js';
-import {type UserDocument} from '../user/user.entity.js';
-import {type UserService} from '../user/user-service.interface.js';
 import {fillDTO} from '../../shared/helpers/common.js';
 import {HttpError} from '../../shared/http-error/http-error.js';
 import {type Logger} from '../../shared/libs/logger/logger.interface.js';
 import {DocumentExistsMiddleware} from '../../shared/middleware/document-exists.middleware.js';
 import {ParseObjectIdMiddleware} from '../../shared/middleware/parse-objectid.middleware.js';
+import {PrivateRouteMiddleware} from '../../shared/middleware/private-route.middleware.js';
 import {ValidateDtoMiddleware} from '../../shared/middleware/validate-dto.middleware.js';
 import {Controller} from '../../shared/rest/controller.abstract.js';
+import {type TokenService} from '../../shared/token/token-service.interface.js';
 import {Component} from '../../shared/types/component.js';
 import {type CityName} from '../../types/entities.js';
 
@@ -29,20 +30,23 @@ export class OfferController extends Controller {
   constructor(
     @inject(Component.Logger) logger: Logger,
     @inject(Component.OfferService) private readonly offerService: OfferService,
-    @inject(Component.UserService) private readonly userService: UserService
+    @inject(Component.TokenService) private readonly tokenService: TokenService
   ) {
     super(logger);
 
     const offerIdMiddleware = new ParseObjectIdMiddleware('offerId');
+    const optionalAuthMiddleware = new PrivateRouteMiddleware(this.tokenService, false);
+    const privateRouteMiddleware = new PrivateRouteMiddleware(this.tokenService);
 
-    this.addRoute({path: '/premium/:cityName', method: 'get', handler: this.getPremiumByCity});
-    this.addRoute({path: '/favorites', method: 'get', handler: this.getFavorites});
+    this.addRoute({path: '/premium/:cityName', method: 'get', handler: this.getPremiumByCity, middlewares: [optionalAuthMiddleware]});
+    this.addRoute({path: '/favorites', method: 'get', handler: this.getFavorites, middlewares: [privateRouteMiddleware]});
     this.addRoute({
       path: '/:offerId/favorite/:status',
       method: 'post',
       handler: this.changeFavoriteStatus,
       middlewares: [
         offerIdMiddleware,
+        privateRouteMiddleware,
         new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')
       ]
     });
@@ -50,7 +54,7 @@ export class OfferController extends Controller {
       path: '/:offerId',
       method: 'get',
       handler: this.show,
-      middlewares: [offerIdMiddleware, new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')]
+      middlewares: [offerIdMiddleware, optionalAuthMiddleware, new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')]
     });
     this.addRoute({
       path: '/:offerId',
@@ -59,6 +63,7 @@ export class OfferController extends Controller {
       middlewares: [
         offerIdMiddleware,
         new ValidateDtoMiddleware(UpdateOfferDto),
+        privateRouteMiddleware,
         new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')
       ]
     });
@@ -66,25 +71,25 @@ export class OfferController extends Controller {
       path: '/:offerId',
       method: 'delete',
       handler: this.delete,
-      middlewares: [offerIdMiddleware, new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')]
+      middlewares: [offerIdMiddleware, privateRouteMiddleware, new DocumentExistsMiddleware(this.offerService, 'Offer', 'offerId', 'offer')]
     });
-    this.addRoute({path: '/', method: 'get', handler: this.index});
-    this.addRoute({path: '/', method: 'post', handler: this.create, middlewares: [new ValidateDtoMiddleware(CreateOfferDto)]});
+    this.addRoute({path: '/', method: 'get', handler: this.index, middlewares: [optionalAuthMiddleware]});
+    this.addRoute({path: '/', method: 'post', handler: this.create, middlewares: [new ValidateDtoMiddleware(CreateOfferDto), privateRouteMiddleware]});
   }
 
   public index = async ({query}: Request, response: Response): Promise<void> => {
     const limit = Number(query.limit) || 60;
-    const offers = await this.offerService.find(limit);
+    const offers = await this.offerService.find(limit, this.getCurrentUserId(response));
     this.ok(response, instanceToPlain(fillDTO(OfferSummaryRdo, offers)));
   };
 
   public show = async (_request: Request<OfferIdParams>, response: Response): Promise<void> => {
-    const offer = response.locals.offer as OfferDocument;
+    const existingOffer = response.locals.offer as OfferDocument;
+    const offer = await this.offerService.findById(existingOffer.id, this.getCurrentUserId(response)) ?? existingOffer;
     this.ok(response, instanceToPlain(fillDTO(OfferDetailRdo, offer)));
   };
 
   public create = async ({body}: Request<object, object, CreateOfferDto>, response: Response): Promise<void> => {
-    const user = await this.resolveUser();
     const dto = body as CreateOfferDto;
     const serviceDto: CreateOfferServiceDto = {
       title: dto.title,
@@ -98,13 +103,14 @@ export class OfferController extends Controller {
       images: dto.images,
       isPremium: dto.isPremium,
       isFavorite: false,
+      favoriteUserIds: [],
       rating: 0,
       housingType: dto.housingType,
       roomsCount: dto.roomsCount,
       guestsCount: dto.guestsCount,
       rentalPrice: dto.rentalPrice,
       amenities: dto.amenities,
-      authorId: user._id,
+      authorId: new Types.ObjectId(this.getRequiredUserId(response)),
       commentsCount: 0
     };
 
@@ -114,6 +120,7 @@ export class OfferController extends Controller {
 
   public update = async ({body}: Request<OfferIdParams, object, UpdateOfferDto>, response: Response): Promise<void> => {
     const existingOffer = response.locals.offer as OfferDocument;
+    this.ensureOfferAuthor(existingOffer, this.getRequiredUserId(response));
 
     const dto = body as UpdateOfferDto;
     const currentCity = existingOffer.city;
@@ -143,40 +150,56 @@ export class OfferController extends Controller {
       };
     }
 
-    const updatedOffer = await this.offerService.updateById(existingOffer.id, serviceDto) ?? existingOffer;
+    const updatedOffer = await this.offerService.updateById(existingOffer.id, serviceDto, this.getRequiredUserId(response)) ?? existingOffer;
 
     this.ok(response, instanceToPlain(fillDTO(OfferDetailRdo, updatedOffer)));
   };
 
   public delete = async ({params}: Request<OfferIdParams>, response: Response): Promise<void> => {
+    this.ensureOfferAuthor(response.locals.offer as OfferDocument, this.getRequiredUserId(response));
     await this.offerService.deleteById(params.offerId);
     this.noContent(response);
   };
 
   public getPremiumByCity = async ({params}: Request<PremiumCityParams>, response: Response): Promise<void> => {
-    const offers = await this.offerService.findPremiumByCity(params.cityName);
+    const offers = await this.offerService.findPremiumByCity(params.cityName, this.getCurrentUserId(response));
     this.ok(response, instanceToPlain(fillDTO(OfferSummaryRdo, offers)));
   };
 
   public getFavorites = async (_request: Request, response: Response): Promise<void> => {
-    const offers = await this.offerService.findFavorites();
+    const offers = await this.offerService.findFavorites(this.getRequiredUserId(response));
     this.ok(response, instanceToPlain(fillDTO(OfferSummaryRdo, offers)));
   };
 
   public changeFavoriteStatus = async ({params}: Request<FavoriteParams>, response: Response): Promise<void> => {
     const isFavorite = params.status === '1';
-    const offer = await this.offerService.updateFavoriteStatus(params.offerId, isFavorite) ?? response.locals.offer as OfferDocument;
+    const offer = await this.offerService.updateFavoriteStatus(params.offerId, this.getRequiredUserId(response), isFavorite)
+      ?? response.locals.offer as OfferDocument;
 
     this.ok(response, instanceToPlain(fillDTO(OfferDetailRdo, offer)));
   };
 
-  private async resolveUser(): Promise<UserDocument> {
-    const user = await this.userService.findFirst();
+  private getCurrentUserId(response: Response): string | undefined {
+    return response.locals.userId as string | undefined;
+  }
 
-    if (!user) {
-      throw new HttpError(StatusCodes.BAD_REQUEST, 'No users available to link offer');
+  private getRequiredUserId(response: Response): string {
+    const userId = this.getCurrentUserId(response);
+
+    if (!userId) {
+      throw new HttpError(StatusCodes.UNAUTHORIZED, 'User is not authenticated');
     }
 
-    return user;
+    return userId;
+  }
+
+  private ensureOfferAuthor(offer: OfferDocument, userId: string): void {
+    const offerAuthorId = typeof offer.authorId === 'object' && offer.authorId !== null && 'id' in offer.authorId
+      ? String(offer.authorId.id)
+      : String(offer.authorId);
+
+    if (offerAuthorId !== userId) {
+      throw new HttpError(StatusCodes.FORBIDDEN, 'You can edit only your own offers');
+    }
   }
 }
